thai-streamlit-app/requirements.txt
streamlit==1.31.1
streamlit-drawable-canvas==0.9.3
easyocr==1.7.1
pillow==10.2.0
pandas==2.2.0
numpy==1.26.3
opencv-python-headless==4.9.0.80
